Added action to add comma as thousands separators in output
Output table includes a toolbar with a single action now to add comma for thousands separators. The formatting is applied for integer and float numbers (short, int, long, float and double). By default, the output keeps as it was previously - without commas. Comma separator mode is applied is individual for every output tab. Toggling this acount doesn't resend the query to server. Result is re-formatted. Shortcut to toggle comma separator is Ctrl+J (or Command+J for MacOS). Note that previously it was possible to have custom DecimalFormat in config file. Now the property is not used.